feat(user, ride, chat, map): Add user's balance to response objects
Adds the balance_in_app field to various response objects across multiple controllers (AuthController, ChatController, MapController, RideController, UserController) to provide users with their current in-app balance.  This field is also added to the swagger documentation.  The database schema was updated to store the balance as an int64 representing the balance in the smallest currency unit.

+++ infra/db/migration/migration_model.go
@@ -36,7 +36,7 @@ type User struct {
 	MomoWalletID       string `gorm:"uniqueIndex"`   // MoMo wallet ID (phone number that is registered with MoMo wallet)
 
 	// New field for storing money received in app
-	BalanceInApp float64 `gorm:"default:0"` // Store balance in cents/smallest currency unit
+	BalanceInApp int64 `gorm:"default:0"` // Store balance in cents/smallest currency unit
 
 	Vehicles          []Vehicle          // One-to-many relationship with Vehicle
 	RatingsReceived   []Rating           `gorm:"foreignKey:RateeID"` // One-to-many relationship with Rating (received)
